backend/requirements.txt
fastapi
uvicorn[standard]
langchain
chromadb
neo4j
openai
python-dotenv
email-validator